requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
ipykernel
requests
pandas
dotenv
minio
pyarrow
python-box
PyYAML
ensure==1.0.2
apache-airflow
apache-airflow-providers-http<5.0.0
apache-airflow-providers-postgres
psycopg2-binary
pytest
pytest-mock
pytest-cov
faker
freezegun
pendulum